Add Become a Starknet Core Developer Talk
Partially fixes #126

Changes made: Adding Become a Starknet Core Developer Conference. 

Reason to add this: This talk by Tomasz Stańczak, founder of Nethermind, provides an in-depth exploration of how to become a core developer within the Starknet ecosystem. It is an excellent resource for developers who want to contribute and don't know the path for it.