Add Kill Switch Contract Example
Partially fixes #126 

Changes made: Add Kill Switch contract repository. 

Reason to add it: This is a simple yet important example of a contract written with Cairo that can be called by any other contract, it's also used by other Starknet Foundation workshops so it's useful to have it as a learning resource here.