Add counter workshop
Partially fixes #126

Changes made: Adding Counter workshop

Reason to add it: This repository offers a hands-on, step-by-step guide to writing smart contracts with Cairo. It’s an ideal starting point for developers new to Cairo and Starknet who want to build basic smart contract functionality, also it's used in Basecamp.